First pass of Users and Groups APIs for the Driver - CRUD for users and groups + membership operations (add/remove)
Currently has separate methods for membership of users and groups (i.e. addUserMember and addGroupMember).
Would be more ideal if you didn't have those but rather:

add
remove
has

These will simply deal with "principals" from which User and Group will extend.  This is a better pattern and easier to use.
It'll be part of the next checkin.